Amounts are double or null

+++ app/src/main/java/com/example/riseep3/ui/componenten/overview/AdjustmentList.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.unit.dp
 
 @Composable
 fun AdjustmentList(
-    adjustments: List<Pair<String, Int>>,
+    adjustments: List<Pair<String, Double>>,
     onEdit: (Int) -> Unit,
     onDelete: (Int) -> Unit
 ) {
